logyca-ai/samples/fastapi_async/requirements.txt
aiohttp==3.9.5
fastapi==0.110.1
gunicorn==21.2.0 # For production
logyca_ai==0.2.1
logyca[api_key_simple_auth]==0.1.18
openai==1.37.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
pydantic==2.6.3
PyJWT==2.8.0
PyMuPDF==1.24.9
pytesseract==0.3.10
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.28.2
starlette==0.37.2
tiktoken==0.7.0
uvicorn==0.30.6